+++ source-code/clj/server_extensions/home/api.clj
@@ -1,3 +1,0 @@
-
-(ns server-extensions.home.api
-    (:require [server-extensions.home.views]))