+++ source-code/clj/server_extensions/home_screen/views.clj
@@ -1,5 +1,5 @@
 
-(ns server-extensions.home.views
+(ns server-extensions.home-screen.views
     (:require [x.server-core.api :as a]))
 
 
@@ -9,7 +9,7 @@
 
 (a/reg-lifecycles!
   ::lifecycles
-  {:on-server-boot [:router/add-route! :home/route
+  {:on-server-boot [:router/add-route! :home-screen/route
                                        {:route-template "/@app-home"
-                                        :client-event   [:home/load!]
+                                        :client-event   [:home-screen/load!]
                                         :restricted?    true}]})