+++ source-code/clj/x/server_ui/body.clj
@@ -5,8 +5,8 @@
 ; Author: bithandshake
 ; Created: 2021.04.19
 ; Description:
-; Version: v0.8.6
-; Compatibility: x4.5.5
+; Version: v0.9.0
+; Compatibility: x4.6.2
 
 
 
@@ -14,7 +14,7 @@
 ;; ----------------------------------------------------------------------------
 
 (ns x.server-ui.body
-    (:require [mid-fruits.candy     :refer [param]]
+    (:require [mid-fruits.candy     :refer [param return]]
               [mid-fruits.string    :as string]
               [mid-fruits.vector    :as vector]
               [x.server-core.api    :as a :refer [cache-control-uri]]
@@ -27,74 +27,33 @@
 
 
 
+;; -- Configuration -----------------------------------------------------------
+;; ----------------------------------------------------------------------------
+
+; @constant (string)
+(def CORE-JS-DIR "/js/core")
+
+
+
 ;; -- Helpers -----------------------------------------------------------------
 ;; ----------------------------------------------------------------------------
 
-(defn- request->core-js-filename
+(defn- core-js-props
   ; WARNING! NON-PUBLIC! DO NOT USE!
   ;
-  ; @param (map) request
+  ; @param (map) body-props
+  ;  {:core-js (string)}
   ;
   ; @example
-  ;  (request->core-js-filename {...})
+  ;  (core-js-props {...})
   ;  =>
-  ;  "app.js"
-  ;
-  ; @return (string)
-  [request]
-  (if-let [core-js-filename (router/request->route-prop request :js)]
-          (string/not-starts-with! core-js-filename  "/")))
-
-(defn- request->core-js-uri-base
-  ; WARNING! NON-PUBLIC! DO NOT USE!
-  ;
-  ; @param (map) request
-  ;
-  ; @example
-  ;  (request->core-js-uri-base {...})
-  ;  =>
-  ;  "/js/core/"
-  ;
-  ; @return (string)
-  [request]
-  (let [core-js-dir @(a/subscribe [:core/get-app-config-item :core-js-dir])]
-       (-> core-js-dir (string/starts-with! "/")
-                       (string/ends-with!   "/"))))
-
-(defn- request->core-js-uri
-  ; WARNING! NON-PUBLIC! DO NOT USE!
-  ;
-  ; @param (map) request
-  ;
-  ; @example
-  ;  (request->core-js-uri-uri {...})
-  ;  =>
-  ;  "/js/core/app.js"
-  ;
-  ; @return (string)
-  [request]
-  (let [core-js-uri-base (request->core-js-uri-base request)
-        core-js-filename (request->core-js-filename request)]
-       (str core-js-uri-base core-js-filename)))
-
-(defn- request->core-js-props
-  ; WARNING! NON-PUBLIC! DO NOT USE!
-  ;
-  ; @param (map) request
-  ;
-  ; @example
-  ;  (request->core-js-uri-props {...})
-  ;  =>
-  ;  {:cache-control? true
-  ;   :uri            "/js/core/app.js"}
+  ;  {:uri "/js/core/app.js"}
   ;
   ; @return (map)
-  ;  {:cache-control? (boolean)
-  ;   :uri (string)}
-  [request]
-  (let [core-js-uri (request->core-js-uri request)]
-       {:cache-control? true
-        :uri            core-js-uri}))
+  ;  {:uri (string)}
+  [{:keys [core-js]}]
+  (let [core-js-uri (str CORE-JS-DIR "/" core-js)]
+       {:uri core-js-uri}))
 
 (defn body<-js-includes
   ; WARNING! NON-PUBLIC! DO NOT USE!
@@ -103,22 +62,24 @@
   ; @param (map) request
   ; @param (map) body-props
   ;  {:app-build (string)(opt)
+  ;   :core-js (string)
   ;   :plugin-js-paths (maps in vector)
-  ;    [{:cache-control? (boolean)(opt)
-  ;       Default: false
+  ;    [{:core-js (string)(opt)
   ;      :uri (string)}]
   ;
   ; @return (hiccup)
-  [body request {:keys [app-build plugin-js-paths]}]
-  (let [core-js-props (request->core-js-props request)
-        js-paths      (vector/cons-item plugin-js-paths core-js-props)]
-       (reduce (fn [body {:keys [cache-control? uri] :as js-props}]
-                   (if cache-control? (let [cache-control-uri (cache-control-uri uri app-build)
-                                            js-props          (assoc js-props :uri cache-control-uri)]
-                                           (conj body (include-js js-props)))
-                                      (conj      body (include-js js-props))))
-               (param body)
-               (param js-paths))))
+  [body request {:keys [app-build core-js plugin-js-paths] :as body-props}]
+  (letfn [(include-js? [js-props] (or (-> js-props :core-js nil?)
+                                      (-> js-props :core-js (= core-js))))
+          (f [body {:keys [uri] :as js-props}]
+             (let [cache-control-uri (cache-control-uri uri app-build)
+                   js-props          (assoc js-props :uri cache-control-uri)]
+                  (if (include-js? js-props)
+                      (conj   body (include-js js-props))
+                      (return body))))]
+         (let [core-js-props (core-js-props body-props)
+               js-paths      (vector/cons-item plugin-js-paths core-js-props)]
+              (reduce f body js-paths))))
 
 
 
@@ -133,11 +94,15 @@
   ;
   ; @return (map)
   ;  {:app-build (string)
+  ;   :core-js (string)
+  ;   :selected-theme (string)
   ;   :shield (hiccup)}
   [request body-props]
   (merge @(a/subscribe [:core/get-app-config])
-          {:app-build (a/app-build)
-           :shield    (app-shield (graphics/loading-animation))}
+          {:app-build      (a/app-build)
+           :core-js        (router/request->route-prop       request :core-js)
+           :selected-theme (user/request->user-settings-item request :selected-theme)
+           :shield         (app-shield (graphics/loading-animation))}
           (param body-props)))
 
 
@@ -150,11 +115,12 @@
   ;
   ; @param (map) request
   ; @param (map) body-props
-  ;  {:shield (hiccup)(opt)}
+  ;  {:selected-theme (string)
+  ;   :shield (hiccup)(opt)}
   ;
   ; @return (hiccup)
-  [request {:keys [shield]}]
-  [:body#x-body-container {:data-theme (user/request->user-settings-item request :selected-theme)}
+  [request {:keys [selected-theme shield]}]
+  [:body#x-body-container {:data-theme selected-theme}
                           (let [csrf-token (force *anti-forgery-token*)]
                                [:div#sente-csrf-token {:data-csrf-token csrf-token}])
                           [:div#x-app-container]
@@ -165,8 +131,7 @@
   ; @param (map)(opt) body-props
   ;  {:app-build (string)(opt)
   ;   :plugin-js-paths (maps in vector)
-  ;    [{:cache-control? (boolean)(opt)
-  ;       Default: false
+  ;    [{:core-js (string)(opt)
   ;      :uri (string)}]
   ;   :shield (hiccup)(opt)}
   ;