+++ source-code/cljs/app_extensions/home/api.cljs
@@ -1,3 +1,0 @@
-
-(ns app-extensions.home.api
-    (:require [app-extensions.home.views]))