+++ source-code/project-emulator/backend.clj
@@ -3,7 +3,7 @@
     (:require [backend.router.api]
               [sente.api]
               [server-extensions.clients.api]
-              [server-extensions.home.api]
+              [server-extensions.home-screen.api]
               [server-extensions.products.api]
               [server-extensions.settings.api]
               [server-extensions.storage.api]