+++ source-code/project-emulator/frontend.cljs
@@ -2,7 +2,7 @@
 (ns frontend
     (:require ;[sente.api]
               [app-extensions.clients.api]
-              [app-extensions.home.api]
+              [app-extensions.home-screen.api]
               [app-extensions.products.api]
               [app-extensions.settings.api]
               [app-extensions.storage.api]